+++ packages/terra-list/src/SingleSelectList.jsx
@@ -17,6 +17,10 @@ const propTypes = {
    */
   hasChevrons: PropTypes.bool,
   /**
+   * Whether or not the child items should be displayed in sections with header and list of elements.
+   */
+  hasSections: PropTypes.bool,
+  /**
    * A callback event that will be triggered when selection state changes.
    */
   onChange: PropTypes.func,
@@ -26,6 +30,7 @@ const defaultProps = {
   children: [],
   isDivided: false,
   hasChevrons: false,
+  hasSections: false,
   onChange: undefined,
 };
 
@@ -34,7 +39,13 @@ class SingleSelectList extends React.Component {
   constructor(props) {
     super(props);
     this.handleOnChange = this.handleOnChange.bind(this);
-    this.state = { selectedIndex: SelectableList.Utils.initialSingleSelectedIndex(this.props.children) };
+    if (!this.props.hasSections) {
+      // if it is not sections, then handle it differently.
+      this.state = { selectedIndex: SelectableList.Utils.initialSingleSelectedIndex(this.props.children) };
+    } else {
+      // hasSections, so handle the index differently.
+      this.state = { selectedIndex: SelectableList.Utils.initialSingleSelectedIndexWithSections(this.props.children) };
+    }
   }
 
   handleOnChange(event, index) {
@@ -48,14 +59,15 @@ class SingleSelectList extends React.Component {
   }
 
   render() {
-    const { children, isDivided, onChange, hasChevrons, ...customProps } = this.props;
+    const { children, isDivided, onChange, hasChevrons, hasSections, ...customProps } = this.props;
     return (
       <SelectableList
         {...customProps}
         isDivided={isDivided}
         onChange={this.handleOnChange}
-        selectedIndexes={[this.state.selectedIndex]}
         hasChevrons={hasChevrons}
+        hasSections={hasSections}
+        selectedIndexes={[this.state.selectedIndex]}
         disableUnselectedItems={false}
       >
         {children}
@@ -67,5 +79,6 @@ class SingleSelectList extends React.Component {
 SingleSelectList.propTypes = propTypes;
 SingleSelectList.defaultProps = defaultProps;
 SingleSelectList.Item = SelectableList.Item;
+SingleSelectList.Section = SelectableList.Section;
 
 export default SingleSelectList;